[CWS] make `packet.filter` fields can be validate when pcap is disabled
